+++ docs/whatsnew/changelog.rst
@@ -29,6 +29,7 @@ Bug fixes
 ~~~~~~~~~
 .. Add here new bug fixes (do not delete this comment)
 
+* #687 fixed.
 
 .. section
 
@@ -42,3 +43,5 @@ Breaking changes
 Deprecations
 ~~~~~~~~~~~~
 .. Add here new deprecations (do not delete this comment)
+
+* `parameters` method of Analysis configurables is now deprecated in favor of `params`.